Implement one type-one module rule for PlonkUp
