+++ src/ZkFold/Base/Protocol/Plonkup/Relation.hs
@@ -18,34 +18,10 @@ import           ZkFold.Base.Algebra.Polynomials.Univariate          (PolyVec, t
 import           ZkFold.Base.Data.Vector                             (Vector, fromVector)
 import           ZkFold.Base.Protocol.Plonkup.LookupConstraint       (LookupConstraint(..))
 import           ZkFold.Base.Protocol.Plonkup.PlonkConstraint        (PlonkConstraint (..), toPlonkConstraint)
+import           ZkFold.Base.Protocol.Plonkup.PlonkupConstraint
 import           ZkFold.Prelude                                      (replicate, length)
 import           ZkFold.Symbolic.Compiler
 import           ZkFold.Symbolic.Compiler.ArithmeticCircuit.Internal
-
-data PlonkupConstraint i a = ConsPlonk (PlonkConstraint i a) | ConsLookup (LookupConstraint i a) | ConsExtra
-
-getPlonkConstraint :: (Eq a, FiniteField a, KnownNat i) => PlonkupConstraint i a -> PlonkConstraint i a
-getPlonkConstraint (ConsPlonk c) = c
-getPlonkConstraint _             = toPlonkConstraint zero
-
-isLookupConstraint :: FiniteField a => PlonkupConstraint i a -> a
-isLookupConstraint (ConsLookup _) = one
-isLookupConstraint _              = zero
-
-getA :: forall a i . (Eq a, FiniteField a, KnownNat i) => PlonkupConstraint i a -> Var (Vector i)
-getA (ConsPlonk c)  = x1 c
-getA (ConsLookup c) = lkVar c
-getA ConsExtra      = x1 (toPlonkConstraint @a zero)
-
-getB :: forall a i . (Eq a, FiniteField a, KnownNat i) => PlonkupConstraint i a -> Var (Vector i)
-getB (ConsPlonk c)  = x2 c
-getB (ConsLookup c) = lkVar c
-getB ConsExtra      = x2 (toPlonkConstraint @a zero)
-
-getC :: forall a i . (Eq a, FiniteField a, KnownNat i) => PlonkupConstraint i a -> Var (Vector i)
-getC (ConsPlonk c)  = x3 c
-getC (ConsLookup c) = lkVar c
-getC ConsExtra      = x3 (toPlonkConstraint @a zero)
 
 -- Here `n` is the total number of constraints, `i` is the number of inputs to the circuit, and `a` is the field type.
 data PlonkupRelation n i a = PlonkupRelation
